@@ -1,51 +1,42 @@
-[tool.poetry]
+[project]
 name = "kwwutils"
-version = "0.1.13"
-description = "General utilities working with Ollama and Langchain"
-authors = ["Kenneth Wong <kennethwork101@gmail.com>"]
+version = "0.1.14"
+description = "Add your description here"
 readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "langchain-chroma>=0.1.4",
+    "langchain>=0.3.13",
+    "langchain-community>=0.3.13",
+    "langchain-ollama>=0.2.2",
+    "transformers>=4.47.1",
+    "pypdf>=5.1.0",
+    "gpt4all>=2.8.2",
+    "sentence-transformers>=3.3.1",
+    "bs4>=0.0.2",
+]
 
-[tool.poetry.dependencies]
-python = "^3.10"
-pypdf = "^5.1.0"
-bs4 = "^0.0.2"
-chromadb = "^0.5.23"
-transformers = "^4.41.0"
-sentence-transformers = "^3.2.0"
-gpt4all = "^2.8.2"
-huggingface = "^0.0.1"
-instructorembedding = "^1.0.1"
-langchain = "^0.3.13"
-tensorflow = "^2.18.0"
-langchain-ollama = "^0.2.2"
-langchain-community = "^0.3.13"
-langchain-huggingface = "^0.1.2"
-langchain-core = "^0.3.28"
-
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.1.1"
-pytest-cov = "^5.0.0"
-pytest-datadir = "^1.5.0"
-pytest-flake8 = "^1.1.1"
-pytest-instafail = "^0.5.0"
-pytest-json = "^0.4.0"
-pytest-leaks = "^0.3.1"
-pytest-pycodestyle = "^2.3.1"
-pytest-redis = "^3.0.2"
-pytest-regressions = "^2.5.0"
-pytest-repeat = "^0.9.3"
-pytest-replay = "^1.5.1"
-pytest-rerunfailures = "^14.0"
-pytest-selenium = "^4.1.0"
-pytest-sugar = "^1.0.0"
-pytest-timeout = "^2.3.1"
-pytest-vagrant = "^2.1.0"
-pytest-vcr = "^1.0.2"
-pytest-xdist = "^3.5.0"
-pytest-xvfb = "^3.0.0"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+[dependency-groups]
+dev = [
+    "pytest>=8.3.4",
+    "pytest-bdd>=7.1.2",
+    "pytest-cov>=5.0.0",
+    "pytest-datadir>=1.5.0",
+    "pytest-flake8>=1.1.1",
+    "pytest-instafail>=0.5.0",
+    "pytest-json>=0.4.0",
+    "pytest-leaks>=0.3.1",
+    "pytest-pycodestyle>=2.3.1",
+    "pytest-redis>=3.0.2",
+    "pytest-regressions>=2.5.0",
+    "pytest-repeat>=0.9.3",
+    "pytest-replay>=1.5.1",
+    "pytest-rerunfailures>=14.0",
+    "pytest-selenium>=4.1.0",
+    "pytest-sugar>=1.0.0",
+    "pytest-timeout>=2.3.1",
+    "pytest-vagrant>=2.1.0",
+    "pytest-vcr>=1.0.2",
+    "pytest-xdist>=3.5.0",
+    "pytest-xvfb>=3.0.0",
+]
